# docs/api-documentation.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1.0/posts": {
      "post": {
        "description": "Creates a new post with the provided content and settings",
        "operationId": "PostController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Post creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePostResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create a new post",
        "tags": [
          "Posts"
        ]
      },
      "get": {
        "description": "Retrieves posts with optional filtering by user ID, hashtag, and pagination",
        "operationId": "PostController_getPosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "minimum": 1,
              "maximum": 10000,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of posts per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter posts by user ID",
            "schema": {
              "example": 42,
              "type": "number"
            }
          },
          {
            "name": "hashtag",
            "required": false,
            "in": "query",
            "description": "Filter posts by hashtag",
            "schema": {
              "example": "#nestjs",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter posts by visibility",
            "schema": {
              "example": "REPLY",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPostsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get posts with optional filters",
        "tags": [
          "Posts"
        ]
      }
    },
    "/api/v1.0/posts/{postId}/like": {
      "post": {
        "description": "Likes a post if not already liked, or unlikes it if already liked",
        "operationId": "PostController_togglePostLike",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "The ID of the post to toggle like",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Like toggled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToggleLikeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid post ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Toggle like on a post",
        "tags": [
          "Posts"
        ]
      }
    },
    "/api/v1.0/posts/{postId}/likers": {
      "get": {
        "description": "Retrieves a paginated list of users who liked the specified post",
        "operationId": "PostController_getPostLikers",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "The ID of the post to get likers for",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of likers per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Likers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLikersResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get list of users who liked a post",
        "tags": [
          "Posts"
        ]
      }
    },
    "/api/v1.0/posts/{postId}/replies": {
      "get": {
        "description": "Retrieves a paginated list of replies to the specified post",
        "operationId": "PostController_getPostReplies",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "The ID of the post to get replies for",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of replies per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Replies retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPostsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get replies to a post",
        "tags": [
          "Posts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hankers",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "error",
              "fail"
            ],
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Invalid input data"
          },
          "error": {
            "type": "object",
            "nullable": true,
            "example": "Bad Request",
            "description": "Optional error details or the type of error"
          }
        },
        "required": [
          "status",
          "message",
          "error"
        ]
      },
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "The textual content of the post",
            "example": "Excited to share my new project today!",
            "maxLength": 500
          },
          "type": {
            "type": "string",
            "description": "The type of post (POST, REPLY, or QUOTE)",
            "enum": [
              "POST",
              "REPLY",
              "QUOTE"
            ],
            "example": "POST"
          },
          "parentId": {
            "type": "number",
            "description": "The ID of the parent post (used when this post is a reply or quote)",
            "example": 42,
            "nullable": true
          },
          "visibility": {
            "type": "string",
            "description": "The visibility level of the post (EVERY_ONE, FOLLOWERS, or MENTIONED)",
            "enum": [
              "EVERY_ONE",
              "FOLLOWERS",
              "MENTIONED"
            ],
            "example": "EVERY_ONE"
          }
        },
        "required": [
          "content",
          "type",
          "visibility"
        ]
      },
...
    }
  }
}